amqpservice/requirements.txt
Django==2.2.4
djangorestframework==3.9.3
psycopg2==2.8.2
celery==4.3
pika==1.1.0
redis==3.3.6
flower==0.9.3
